requirements.txt
streamlit
face_recognition
opencv-python
numpy
pandas
Pillow


--------------------------------------------------------------------------------------------------------------------


---

# 🧾 **requirements.txt (Final Version)**

```text
streamlit
opencv-python
numpy
pandas
Pillow
face_recognition
dlib==19.24.2
